third_party/WebKit/LayoutTests/batterystatus/updates-expected.txt

CONSOLE MESSAGE: line 40: Uncaught TypeError: Cannot call method 'addEventListener' of undefined
Tests that updates to the battery event causes new events to fire.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL successfullyParsed should be true (of type boolean). Was undefined (of type undefined).

TEST COMPLETE
